pyproject.toml
[tool.poetry]
name = "tap-hubspot"
version = "0.0.2"
description = "a Singer tap for extracting data from Hubspot"
authors = ["Dialogue <hackers@dialogue.co>"]

[tool.poetry.dependencies]
python = "^3.7"
attrs = "^19.1"
singer-python = "^5.8"
requests = "^2.22"
requests_mock = "^1.6"
nose = "^1.3"
structlog = "^19.1"
"dialogue.logging" = "^2.6.1"

[tool.poetry.dev-dependencies]
pylama = "^7.7"
isort = "^4.3"
black = "19.3b0"

[[tool.poetry.source]]
name = "dialogue"
url = "https://pypi.fury.io/dialogue"

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"

[tool.black]
exclude = '.vscode|.eggs|.venv'

[tool.poetry.scripts]
tap-hubspot = "tap_hubspot:main"
